Move extensions to another package
