+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/code/fragment/CodeEditorFragment.kt
@@ -5,8 +5,8 @@ import android.view.View
 import androidx.fragment.app.Fragment
 import by.kirich1409.viewbindingdelegate.viewBinding
 import org.hyperskill.app.android.R
-import org.hyperskill.app.android.code.model.ProgrammingLanguage
-import org.hyperskill.app.android.code.model.extensionForLanguage
+import org.hyperskill.app.android.code.presentation.model.ProgrammingLanguage
+import org.hyperskill.app.android.code.presentation.model.extensionForLanguage
 import org.hyperskill.app.android.databinding.LayoutCodeEditorTestBinding
 
 // TODO: Remove before pushing to develop
