+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/core/extensions/RxUtil.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.app.android.util
+package org.hyperskill.app.android.core.extensions
 
 import io.reactivex.Completable
 import io.reactivex.Flowable
@@ -7,7 +7,7 @@ import io.reactivex.ObservableEmitter
 import io.reactivex.Single
 import io.reactivex.functions.BiFunction
 import io.reactivex.rxkotlin.zipWith
-import org.hyperskill.app.android.util.rx.ObservableReduceMap
+import org.hyperskill.app.android.core.extensions.rx.ObservableReduceMap
 import java.util.concurrent.TimeUnit
 import kotlin.math.pow
